refactored unit-tests for the alerts-warnings page
