+++ test/data/nonLocationGetJson.json
@@ -1,0 +1,57 @@
+{
+    "authenticationResultCode": "ValidCredentials",
+    "brandLogoUri": "http://dev.virtualearth.net/Branding/logo_powered_by.png",
+    "copyright": "Copyright",
+    "resourceSets": [
+      {
+        "estimatedTotal": 2,
+        "resources": [
+          {
+            "__type": "Location:http://schemas.microsoft.com/search/local/ws/rest/v1",
+            "bbox": [
+              -20.939201,
+              167.21288,
+              -20.88126,
+              167.29555
+            ],
+            "name": "Wé, New Caledonia",
+            "point": {
+              "type": "Point",
+              "coordinates": [
+                -20.91023,
+                167.25421
+              ]
+            },
+            "address": {
+              "countryRegion": "New Caledonia",
+              "formattedAddress": "Wé, New Caledonia",
+              "locality": "Wé",
+              "countryRegionIso2": "NC"
+            },
+            "confidence": "Low",
+            "entityType": "PopulatedPlace",
+            "geocodePoints": [
+              {
+                "type": "Point",
+                "coordinates": [
+                  -20.91023,
+                  167.25421
+                ],
+                "calculationMethod": "Rooftop",
+                "usageTypes": [
+                  "Display"
+                ]
+              }
+            ],
+            "matchCodes": [
+              "Good"
+            ]
+          }
+        ]
+      }
+    ],
+    "statusCode": 200,
+    "tatusDescription": "OK",
+    "traceId": "trace-id"
+  }
+
