--- test/routes/alerts-and-warnings-3.js
@@ -1,126 +1,0 @@
-'use strict'
-
-const Hapi = require('@hapi/hapi')
-const Lab = require('@hapi/lab')
-const Code = require('@hapi/code')
-const sinon = require('sinon')
-const lab = exports.lab = Lab.script()
-const data = require('../data')
-
-lab.experiment('Test - alerts - warnings', () => {
-  let sandbox
-  let server
-
-  // Use a Sinon sandbox to manage spies, stubs and mocks for each test.
-  lab.beforeEach(async () => {
-    delete require.cache[require.resolve('../../server/util.js')]
-    delete require.cache[require.resolve('../../server/services/location.js')]
-    delete require.cache[require.resolve('../../server/services/flood.js')]
-    delete require.cache[require.resolve('../../server/routes/location.js')]
-    delete require.cache[require.resolve('../../server/routes/alerts-and-warnings.js')]
-    sandbox = await sinon.createSandbox()
-    server = Hapi.server({
-      port: 3000,
-      host: 'localhost'
-    })
-  })
-
-  lab.afterEach(async () => {
-    await server.stop()
-    await sandbox.restore()
-  })
-
-  lab.test('GET /alerts-and-warnings with query parameters of WA4 1HT', async () => {
-    const floodService = require('../../server/services/flood')
-
-    const fakeIsEngland = () => {
-      return { is_england: true }
-    }
-    // const fakePlaceData = () => {}
-    const fakeFloodsData = () => data.floodsByPostCode
-
-    const fakeStationsData = () => []
-    const fakeImpactsData = () => []
-
-    sandbox.stub(floodService, 'getIsEngland').callsFake(fakeIsEngland)
-    sandbox.stub(floodService, 'getFloodsWithin').callsFake(fakeFloodsData)
-    sandbox.stub(floodService, 'getStationsWithin').callsFake(fakeStationsData)
-    sandbox.stub(floodService, 'getImpactsWithin').callsFake(fakeImpactsData)
-
-    const fakeGetJson = () => {
-      return {
-        authenticationResultCode: 'ValidCredentials',
-        brandLogoUri: 'http://dev.virtualearth.net/Branding/logo_powered_by.png',
-        copyright: 'Copyright',
-        resourceSets: [
-          {
-            estimatedTotal: 1,
-            resources: [
-              {
-                __type: 'Location:http://schemas.microsoft.com/search/local/ws/rest/v1',
-                bbox: [53.37659671137902, -2.579770244840125, 53.384322146520375, -2.5625034058876555],
-                name: 'WA4 1HT, Warrington, Warrington',
-                point: {
-                  type: 'Point',
-                  coordinates: [53.3804594289497, -2.57113682536389]
-                },
-                address: {
-                  adminDistrict: 'England',
-                  adminDistrict2: 'Warrington',
-                  countryRegion: 'United Kingdom',
-                  formattedAddress: 'WA4 1HT, Warrington, Warrington',
-                  locality: 'Warrington',
-                  poastalCode: 'WA4 1HT',
-                  countryRegionIso2: 'GB'
-                },
-                confidence: 'Medium',
-                entityType: 'PopulatedPlace',
-                geocodePoints: [
-                  {
-                    type: 'Point',
-                    coordinates: [53.3804594289497, -2.57113682536389],
-                    calculationMethod: 'None',
-                    usageTypes: ['Display']
-                  }
-                ],
-                matchCodes: ['Good']
-              }
-            ]
-          }
-        ],
-        statusCode: 200,
-        tatusDescription: 'OK',
-        traceId: 'trace-id'
-      }
-    }
-
-    const util = require('../../server/util')
-    sandbox.stub(util, 'getJson').callsFake(fakeGetJson)
-
-    const warningsPlugin = {
-      plugin: {
-        name: 'warnings',
-        register: (server, options) => {
-          server.route(require('../../server/routes/alerts-and-warnings'))
-        }
-      }
-    }
-
-    await server.register(require('../../server/plugins/views'))
-    await server.register(require('../../server/plugins/session'))
-    await server.register(warningsPlugin)
-
-    await server.initialize()
-    const options = {
-      method: 'GET',
-      url: '/alerts-and-warnings?q=WA4%201HT'
-    }
-
-    const response = await server.inject(options)
-
-    Code.expect(response.payload).to.contain('4 results')
-    Code.expect(response.payload).to.contain('4 flood alerts')
-    Code.expect(response.payload).to.contain('<a href="/target-area/013WAFLM">Lower River Mersey including Warrington, Runcorn and Lymm areas</a>')
-    Code.expect(response.statusCode).to.equal(200)
-  })
-})
