--- test/routes/alerts-and-warnings-Scotland.js
@@ -1,114 +1,0 @@
-'use strict'
-
-const Hapi = require('@hapi/hapi')
-const Lab = require('@hapi/lab')
-const Code = require('@hapi/code')
-const sinon = require('sinon')
-const lab = exports.lab = Lab.script()
-
-lab.experiment('Test - alerts - warnings', () => {
-  let sandbox
-  let server
-
-  // Use a Sinon sandbox to manage spies, stubs and mocks for each test.
-  lab.beforeEach(async () => {
-    delete require.cache[require.resolve('../../server/util.js')]
-    delete require.cache[require.resolve('../../server/services/location.js')]
-    delete require.cache[require.resolve('../../server/services/flood.js')]
-    delete require.cache[require.resolve('../../server/routes/location.js')]
-    delete require.cache[require.resolve('../../server/routes/alerts-and-warnings.js')]
-    sandbox = await sinon.createSandbox()
-    server = Hapi.server({
-      port: 3000,
-      host: 'localhost'
-    })
-  })
-
-  lab.afterEach(async () => {
-    await server.stop()
-    await sandbox.restore()
-  })
-
-  lab.test('GET /alerts-and-warnings with query parameters of Kinghorn, Scotland', async () => {
-    const floodService = require('../../server/services/flood')
-
-    const fakeIsEngland = () => {
-      return { is_england: false }
-    }
-
-    sandbox.stub(floodService, 'getIsEngland').callsFake(fakeIsEngland)
-
-    const fakeGetJson = () => {
-      return {
-        authenticationResultCode: 'ValidCredentials',
-        brandLogoUri: 'http://dev.virtualearth.net/Branding/logo_powered_by.png',
-        copyright: 'Copyright',
-        resourceSets: [
-          {
-            estimatedTotal: 1,
-            resources: [
-              {
-                __type: 'Location:http://schemas.microsoft.com/search/local/ws/rest/v1',
-                bbox: [56.061851501464844, -3.1849100589752197, 56.07693099975586, -3.16759991645813],
-                name: 'Kinghorn, Fife',
-                point: {
-                  type: 'Point',
-                  coordinates: [56.0693359375, -3.1757969856262207]
-                },
-                address: {
-                  adminDistrict: 'Scotland',
-                  adminDistrict2: 'Fife',
-                  countryRegion: 'United Kingdom',
-                  formattedAddress: 'Kinghorn, Fife',
-                  locality: 'Kinghorn',
-                  countryRegionIso2: 'GB'
-                },
-                confidence: 'High',
-                entityType: 'PopulatedPlace',
-                geocodePoints: [
-                  {
-                    type: 'Point',
-                    coordinates: [56.0693359375, -3.1757969856262207],
-                    calculationMethod: 'Rooftop',
-                    usageTypes: ['Display']
-                  }
-                ],
-                matchCodes: ['Good']
-              }
-            ]
-          }
-        ],
-        statusCode: 200,
-        tatusDescription: 'OK',
-        traceId: 'trace-id'
-      }
-    }
-
-    const util = require('../../server/util')
-    sandbox.stub(util, 'getJson').callsFake(fakeGetJson)
-
-    const warningsPlugin = {
-      plugin: {
-        name: 'warnings',
-        register: (server, options) => {
-          server.route(require('../../server/routes/alerts-and-warnings'))
-        }
-      }
-    }
-
-    await server.register(require('../../server/plugins/views'))
-    await server.register(require('../../server/plugins/session'))
-    await server.register(warningsPlugin)
-
-    await server.initialize()
-    const options = {
-      method: 'GET',
-      url: '/alerts-and-warnings?q=kinghorn'
-    }
-
-    const response = await server.inject(options)
-
-    Code.expect(response.payload).to.contain('<h1 class="govuk-heading-xl">This service provides flood warning information for England only</h1>')
-    Code.expect(response.statusCode).to.equal(200)
-  })
-})
