+++ test/routes/alerts-and-warnings.js
@@ -5,8 +5,9 @@ const Lab = require('@hapi/lab')
 const Code = require('@hapi/code')
 const sinon = require('sinon')
 const lab = exports.lab = Lab.script()
+const data = require('../data')
 
-lab.experiment('Test - alerts - warnings', () => {
+lab.experiment('Test - /alerts-warnings', () => {
   let sandbox
   let server
 
@@ -29,13 +30,13 @@ lab.experiment('Test - alerts - warnings', () => {
     await sandbox.restore()
   })
 
-  lab.test('GET /alerts-and-warnings with query parameters and no warnings or alerts', async () => {
+  lab.test('GET /alerts-and-warnings with query parameters of Warrington and no warnings or alerts', async () => {
     const floodService = require('../../server/services/flood')
 
     const fakeIsEngland = () => {
       return { is_england: true }
     }
-    // const fakePlaceData = () => {}
+
     const fakeFloodsData = () => {
       return { floods: [] }
     }
@@ -47,51 +48,7 @@ lab.experiment('Test - alerts - warnings', () => {
     sandbox.stub(floodService, 'getStationsWithin').callsFake(fakeStationsData)
     sandbox.stub(floodService, 'getImpactsWithin').callsFake(fakeImpactsData)
 
-    const fakeGetJson = () => {
-      return {
-        authenticationResultCode: 'ValidCredentials',
-        brandLogoUri: 'http://dev.virtualearth.net/Branding/logo_powered_by.png',
-        copyright: 'Copyright',
-        resourceSets: [
-          {
-            estimatedTotal: 1,
-            resources: [
-              {
-                __type: 'Location:http://schemas.microsoft.com/search/local/ws/rest/v1',
-                bbox: [53.367538452148437, -2.6395580768585205, 53.420841217041016, -2.5353000164031982],
-                name: 'Warrington, Warrington',
-                point: {
-                  type: 'Point',
-                  coordinates: [53.393871307373047, -2.5893499851226807]
-                },
-                address: {
-                  adminDistrict: 'England',
-                  adminDistrict2: 'Warrington',
-                  countryRegion: 'United Kingdom',
-                  formattedAddress: 'Warrington, Warrington',
-                  locality: 'Warrington',
-                  countryRegionIso2: 'GB'
-                },
-                confidence: 'Medium',
-                entityType: 'PopulatedPlace',
-                geocodePoints: [
-                  {
-                    type: 'Point',
-                    coordinates: [53.393871307373047, -2.5893499851226807],
-                    calculationMethod: 'Rooftop',
-                    usageTypes: ['Display']
-                  }
-                ],
-                matchCodes: ['Good']
-              }
-            ]
-          }
-        ],
-        statusCode: 200,
-        tatusDescription: 'OK',
-        traceId: 'trace-id'
-      }
-    }
+    const fakeGetJson = () => data.warringtonGetJson
 
     const util = require('../../server/util')
     sandbox.stub(util, 'getJson').callsFake(fakeGetJson)
@@ -122,52 +79,7 @@ lab.experiment('Test - alerts - warnings', () => {
     Code.expect(response.statusCode).to.equal(200)
   })
   lab.test('GET /alerts-and-warnings TYPO or non location "afdv vdaf adfv  fda" ', async () => {
-    const fakeGetJson = () => {
-      return {
-        authenticationResultCode: 'ValidCredentials',
-        brandLogoUri: 'http://dev.virtualearth.net/Branding/logo_powered_by.png',
-        copyright: 'Copyright',
-        resourceSets: [
-          {
-            estimatedTotal: 2,
-            resources: [
-              {
-                __type: 'Location:http://schemas.microsoft.com/search/local/ws/rest/v1',
-                bbox: [-20.939201,
-                  167.21288,
-                  -20.88126,
-                  167.29555],
-                name: 'Wé, New Caledonia',
-                point: {
-                  type: 'Point',
-                  coordinates: [-20.91023, 167.25421]
-                },
-                address: {
-                  countryRegion: 'New Caledonia',
-                  formattedAddress: 'Wé, New Caledonia',
-                  locality: 'Wé',
-                  countryRegionIso2: 'NC'
-                },
-                confidence: 'Low',
-                entityType: 'PopulatedPlace',
-                geocodePoints: [
-                  {
-                    type: 'Point',
-                    coordinates: [-20.91023, 167.25421],
-                    calculationMethod: 'Rooftop',
-                    usageTypes: ['Display']
-                  }
-                ],
-                matchCodes: ['Good']
-              }
-            ]
-          }
-        ],
-        statusCode: 200,
-        tatusDescription: 'OK',
-        traceId: 'trace-id'
-      }
-    }
+    const fakeGetJson = () => data.nonLocationGetJson
 
     const util = require('../../server/util')
     sandbox.stub(util, 'getJson').callsFake(fakeGetJson)
@@ -195,6 +107,145 @@ lab.experiment('Test - alerts - warnings', () => {
 
     Code.expect(response.payload).to.contain('No flood alerts or warnings near this location.')
     Code.expect(response.payload).to.contain('0 results')
+    Code.expect(response.statusCode).to.equal(200)
+  })
+  lab.test('GET /alerts-and-warnings with query parameters of Kinghorn, Scotland', async () => {
+    const floodService = require('../../server/services/flood')
+
+    const fakeIsEngland = () => {
+      return { is_england: false }
+    }
+
+    sandbox.stub(floodService, 'getIsEngland').callsFake(fakeIsEngland)
+
+    const fakeGetJson = () => data.scotlandGetJson
+
+    const util = require('../../server/util')
+    sandbox.stub(util, 'getJson').callsFake(fakeGetJson)
+
+    const warningsPlugin = {
+      plugin: {
+        name: 'warnings',
+        register: (server, options) => {
+          server.route(require('../../server/routes/alerts-and-warnings'))
+        }
+      }
+    }
+
+    await server.register(require('../../server/plugins/views'))
+    await server.register(require('../../server/plugins/session'))
+    await server.register(warningsPlugin)
+
+    await server.initialize()
+    const options = {
+      method: 'GET',
+      url: '/alerts-and-warnings?q=kinghorn'
+    }
+
+    const response = await server.inject(options)
+
+    Code.expect(response.payload).to.contain('<h1 class="govuk-heading-xl">This service provides flood warning information for England only</h1>')
+    Code.expect(response.statusCode).to.equal(200)
+  })
+  lab.test('GET /alerts-and-warnings with query parameters, show alert, warnings and severe', async () => {
+    const floodService = require('../../server/services/flood')
+
+    const fakeIsEngland = () => {
+      return { is_england: true }
+    }
+
+    const fakeFloodsData = () => data.fakeFloodsData
+
+    const fakeStationsData = () => []
+    const fakeImpactsData = () => []
+
+    sandbox.stub(floodService, 'getIsEngland').callsFake(fakeIsEngland)
+    sandbox.stub(floodService, 'getFloodsWithin').callsFake(fakeFloodsData)
+    sandbox.stub(floodService, 'getStationsWithin').callsFake(fakeStationsData)
+    sandbox.stub(floodService, 'getImpactsWithin').callsFake(fakeImpactsData)
+
+    const fakeGetJson = () => data.warringtonGetJson
+
+    const util = require('../../server/util')
+    sandbox.stub(util, 'getJson').callsFake(fakeGetJson)
+
+    const warningsPlugin = {
+      plugin: {
+        name: 'warnings',
+        register: (server, options) => {
+          server.route(require('../../server/routes/alerts-and-warnings'))
+        }
+      }
+    }
+
+    await server.register(require('../../server/plugins/views'))
+    await server.register(require('../../server/plugins/session'))
+    await server.register(warningsPlugin)
+
+    await server.initialize()
+    const options = {
+      method: 'GET',
+      url: '/alerts-and-warnings?q=Warrington'
+    }
+
+    const response = await server.inject(options)
+
+    Code.expect(response.payload).to.contain('6 results')
+    Code.expect(response.payload).to.contain('1 flood warning')
+    Code.expect(response.payload).to.contain('1 severe flood warning')
+    Code.expect(response.payload).to.contain('3 flood alerts')
+    Code.expect(response.payload).to.contain('1 flood warning removed')
+    Code.expect(response.payload).to.contain('<a href="/target-area/013WAFGL">River Glaze catchment including Leigh and East Wigan</a>')
+    Code.expect(response.payload).to.contain('<a href="/target-area/013FWFCH29">Wider area at risk from Sankey Brook at Dallam</a>')
+    Code.expect(response.payload).to.contain('<a href="/target-area/013WAFDI">River Ditton catchment including areas around Huyton-with-Roby and Widnes</a>')
+    Code.expect(response.statusCode).to.equal(200)
+  })
+  lab.test('GET /alerts-and-warnings with query parameters of WA4 1HT', async () => {
+    const floodService = require('../../server/services/flood')
+
+    const fakeIsEngland = () => {
+      return { is_england: true }
+    }
+
+    const fakeFloodsData = () => data.floodsByPostCode
+
+    const fakeStationsData = () => []
+    const fakeImpactsData = () => []
+
+    sandbox.stub(floodService, 'getIsEngland').callsFake(fakeIsEngland)
+    sandbox.stub(floodService, 'getFloodsWithin').callsFake(fakeFloodsData)
+    sandbox.stub(floodService, 'getStationsWithin').callsFake(fakeStationsData)
+    sandbox.stub(floodService, 'getImpactsWithin').callsFake(fakeImpactsData)
+
+    const fakeGetJson = () => data.warringtonGetJson
+
+    const util = require('../../server/util')
+    sandbox.stub(util, 'getJson').callsFake(fakeGetJson)
+
+    const warningsPlugin = {
+      plugin: {
+        name: 'warnings',
+        register: (server, options) => {
+          server.route(require('../../server/routes/alerts-and-warnings'))
+        }
+      }
+    }
+
+    await server.register(require('../../server/plugins/views'))
+    await server.register(require('../../server/plugins/session'))
+    await server.register(warningsPlugin)
+
+    await server.initialize()
+    const options = {
+      method: 'GET',
+      url: '/alerts-and-warnings?q=WA4%201HT'
+    }
+
+    const response = await server.inject(options)
+
+    Code.expect(response.payload).to.contain('4 results')
+    Code.expect(response.payload).to.contain('4 flood alerts')
+    Code.expect(response.payload).to.contain('<a href="/target-area/013WAFLM">Lower River Mersey including Warrington, Runcorn and Lymm areas</a>')
     Code.expect(response.statusCode).to.equal(200)
   })
 })
